+++ projects/ssr-tests/src/ssr-testing.spec.ts
@@ -1,5 +1,6 @@
 import { Server } from 'http';
 import { expectLogMessages } from './matchers/matchers';
+import * as HttpUtils from './utils/http.utils';
 import * as LogUtils from './utils/log.utils';
 import * as ProxyUtils from './utils/proxy.utils';
 import * as SsrUtils from './utils/ssr.utils';
@@ -30,7 +31,7 @@ describe('SSR E2E', () => {
           target: BACKEND_BASE_URL,
         });
         const response: any =
-          await ProxyUtils.sendRequestToSsrServer(REQUEST_PATH);
+          await HttpUtils.sendRequestToSsrServer(REQUEST_PATH);
         expect(response.statusCode).toEqual(200);
 
         expectLogMessages().toContainLogs([
@@ -43,7 +44,7 @@ describe('SSR E2E', () => {
         backendProxy = await ProxyUtils.startBackendProxyServer({
           target: BACKEND_BASE_URL,
         });
-        const response = await ProxyUtils.sendRequestToSsrServer(
+        const response = await HttpUtils.sendRequestToSsrServer(
           REQUEST_PATH + 'not-existing-page'
         );
         expect(response.statusCode).toEqual(404);
@@ -58,7 +59,7 @@ describe('SSR E2E', () => {
             }
           },
         });
-        const response = await ProxyUtils.sendRequestToSsrServer(REQUEST_PATH);
+        const response = await HttpUtils.sendRequestToSsrServer(REQUEST_PATH);
         expect(response.statusCode).toEqual(404);
       });
 
@@ -71,7 +72,7 @@ describe('SSR E2E', () => {
             }
           },
         });
-        const response = await ProxyUtils.sendRequestToSsrServer(REQUEST_PATH);
+        const response = await HttpUtils.sendRequestToSsrServer(REQUEST_PATH);
         expect(response.statusCode).toEqual(500);
       });
     });
@@ -85,8 +86,8 @@ describe('SSR E2E', () => {
         backendProxy = await ProxyUtils.startBackendProxyServer({
           target: BACKEND_BASE_URL,
         });
-        let response: ProxyUtils.SsrResponse;
-        response = await ProxyUtils.sendRequestToSsrServer(REQUEST_PATH);
+        let response: HttpUtils.SsrResponse;
+        response = await HttpUtils.sendRequestToSsrServer(REQUEST_PATH);
         expect(response.statusCode).toEqual(200);
 
         expectLogMessages().toContainLogs([
@@ -94,7 +95,7 @@ describe('SSR E2E', () => {
           `Request is waiting for the SSR rendering to complete (${REQUEST_PATH})`,
         ]);
 
-        response = await ProxyUtils.sendRequestToSsrServer(REQUEST_PATH);
+        response = await HttpUtils.sendRequestToSsrServer(REQUEST_PATH);
         expect(response.statusCode).toEqual(200);
         expectLogMessages().toContain(`Render from cache (${REQUEST_PATH})`);
       });
@@ -108,12 +109,11 @@ describe('SSR E2E', () => {
             }
           },
         });
-        let response: ProxyUtils.SsrResponse;
-        response = await ProxyUtils.sendRequestToSsrServer(REQUEST_PATH);
+        let response: HttpUtils.SsrResponse;
+        response = await HttpUtils.sendRequestToSsrServer(REQUEST_PATH);
         expect(response.statusCode).toEqual(404);
-        LogUtils.clearSsrLogFile();
 
-        response = await ProxyUtils.sendRequestToSsrServer(REQUEST_PATH);
+        response = await HttpUtils.sendRequestToSsrServer(REQUEST_PATH);
         expect(response.statusCode).toEqual(404);
         expectLogMessages().not.toContain(
           `Render from cache (${REQUEST_PATH})`
