+++ projects/ssr-tests/src/utils/proxy.utils.ts
@@ -6,29 +6,9 @@
  */
 
 import * as http from 'http';
-import httpProxy, { ProxyResCallback } from 'http-proxy';
+import httpProxy from 'http-proxy';
 
-const proxy = httpProxy.createProxyServer({
-  secure: false,
-  // selfHandleResponse: true,
-});
-
-/**
- * Default settings to send http requests.
- */
-const REQUEST_OPTIONS = {
-  host: 'localhost',
-  port: 4000,
-};
-
-/**
- * Response from SSR server.
- */
-export interface SsrResponse {
-  statusCode: number | undefined;
-  headers: http.IncomingHttpHeaders;
-  body: string;
-}
+let proxy: httpProxy;
 
 /**
  * Options to start a proxy server.
@@ -45,7 +25,7 @@ interface ProxyOptions {
   /**
    * Callback to be executed if the request to the target got a response.
    */
-  callback?: ProxyResCallback;
+  callback?: httpProxy.ProxyResCallback;
 }
 
 /**
@@ -54,6 +34,9 @@ interface ProxyOptions {
 export async function startBackendProxyServer(
   options: ProxyOptions
 ): Promise<http.Server> {
+  proxy = httpProxy.createProxyServer({
+    secure: false,
+  });
   return new Promise((resolve) => {
     const server = http.createServer((req, res) => {
       const forwardRequest = () =>
@@ -76,32 +59,5 @@ export async function startBackendProxyServer(
     server.listen(9002, () => {
       resolve(server);
     });
-  });
-}
-
-/**
- * Send an http GET request with given URL to the SSR server.
- */
-export async function sendRequestToSsrServer(
-  path: string
-): Promise<SsrResponse> {
-  return new Promise((resolve, reject) => {
-    http
-      .get({ ...REQUEST_OPTIONS, path }, (res) => {
-        let body = '';
-        res.on('data', (chunk) => {
-          body += chunk;
-        });
-        res.on('end', () => {
-          resolve({
-            statusCode: res.statusCode,
-            headers: res.headers,
-            body,
-          });
-        });
-      })
-      .on('error', (e: Error) => {
-        reject(e);
-      });
   });
 }
